+++ docs/recipes/automation/tp-setup/bootstrap/case/create_global_config.py
@@ -20,6 +20,7 @@ if __name__ == "__main__":
         po_dp_config = PageObjectDataPlaneConfiguration(page)
         # config global dataplane
         po_dp_config.o11y_config_dataplane_resource(ENV.TP_AUTO_DP_NAME_GLOBAL)
+        po_dp_config.o11y_config_activation(ENV.TP_AUTO_DP_NAME_GLOBAL)
 
         po_auth.logout()
     except Exception as e: