+++ docs/recipes/automation/tp-setup/bootstrap/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [10/04/2025 22:17]
+### Added
+- Support for creating OAuth Token and saving it to kubernetes secret.
+### Fixed
+- Link k8s dataplane activation url to global.
+
 ## [09/27/2025 21:30]
 ### Fixed
 - Fixed deploy flogo issue after flogo UI is changed